Sync geth nits (#585)
* Change feehistory input type from int to uint64

* Add err handling for UnpackLog with zero topics in log

+++ eth/gasprice/feehistory_test.go
@@ -44,9 +44,9 @@ import (
 func TestFeeHistory(t *testing.T) {
 	var cases = []struct {
 		pending      bool
-		maxCallBlock int
+		maxCallBlock uint64
 		maxBlock     int
-		count        int
+		count        uint64
 		last         rpc.BlockNumber
 		percent      []float64
 		expFirst     uint64
